+++ CHANGELOG.rst
@@ -1,10 +1,11 @@
 Changelog
 =========
 
-Current (2026-09-11)
+Current (2026-09-14)
 --------------------
 
 * Fixed typing to follow modern practice
+* Updated waveform yielding to use a waveform record
 
 
 0.2.0 (2026-09-09)